+++ changelog.md
@@ -1,7 +1,7 @@
 # v1.0.1
 
  * Android support!
- * Added volume slider for SFX
+ * Added volume slider for SFX (only in mod settings)
  * You can select between two algorithms for character selection (randomized or sequential)
  * Added tags for indexer
  * Normalized Sound for Sprite Pack 1